engine: extract esm/wtml.js
